requirements.txt
pdfplumber
sentence-transformers
rank-bm25
numpy
tqdm
regex
scikit-learn
faiss-cpu
python-dotenv
